log: move TraceLogger to its own module
